#3 - Move ApplicationTest from Unit to Integration directory
