+++ phpunit.xml
@@ -19,5 +19,9 @@
         <testsuite name="unit">
             <directory>./tests/Unit</directory>
         </testsuite>
+
+        <testsuite name="integration">
+            <directory>./tests/Integration</directory>
+        </testsuite>
     </testsuites>
 </phpunit>
